-Renamed http.py to http_status.py in order to avoid conflicts with http package used for debugging purposes on Python3
